test: [metrics] rename fixtures files
