+++ src/tests/store/metrics-store/helpers/fixtures/convert-beholder-to-chart-fixtures.js
